+++ aux_data/lsst_i_star.asc
