+++ aux_data/lsst_i_star_zf5.asc
