+++ aux_data/lsst_r_star.asc
