+++ aux_data/lsst_r_star_zf5.asc
