+++ aux_data/lsst_y_star_zf5.asc
