+++ aux_data/lsst_z_star.asc
