+++ aux_data/lsst_z_star_zf5.asc
