+++ data.cfg
@@ -36,18 +36,18 @@ footprint:
 
 magstar_file:
     DES:
-        filter_i: './aux/mstar_i_DES.asc'
-        filter_z: './aux/mstar_z_DES.asc'
+        filter_i: './aux_data/mstar_i_DES.asc'
+        filter_z: './aux_data/mstar_z_DES.asc'
     DC2:
-        filter_r: './aux/lsst_r_star_zf5.asc'
-        filter_i: './aux/lsst_i_star_zf5.asc'
-        filter_z: './aux/lsst_z_star_zf5.asc'
-        filter_y: './aux/lsst_y_star_zf5.asc'
+        filter_r: './aux_data/lsst_r_star_zf5.asc'
+        filter_i: './aux_data/lsst_i_star_zf5.asc'
+        filter_z: './aux_data/lsst_z_star_zf5.asc'
+        filter_y: './aux_data/lsst_y_star_zf5.asc'
     DC2_test:
-        filter_r: './aux/lsst_r_star_zf5.asc'
-        filter_i: './aux/lsst_i_star_zf5.asc'
-        filter_z: './aux/lsst_z_star_zf5.asc'
-        filter_y: './aux/lsst_y_star_zf5.asc'
+        filter_r: './aux_data/lsst_r_star_zf5.asc'
+        filter_i: './aux_data/lsst_i_star_zf5.asc'
+        filter_z: './aux_data/lsst_z_star_zf5.asc'
+        filter_y: './aux_data/lsst_y_star_zf5.asc'
 
 
 # PHOTO Zs
