+++ wazp/__init__.py
@@ -1,0 +1,3 @@
+""" WaZP is a code to detect cluster based on photo-zs and a wavelet decomposition."""
+
+__version__ = "0.1.0"
