+++ wazp/detection.py
