+++ wazp/pmem.py
